requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
faiss-cpu
fastapi
uvicorn
pygame==2.5.2